Ensures integer pixel values and adds getters for subpixel values
Newer browsers are returning pixel values as doubles instead of
integers especially when using zooming. This is causing problems where
existing code is assuming integer values are returned and getting
doubles instead.

To fix this this patch introduces subpixel variants of most DOM
methods dealing with pixels and makes sure the existing methods
are returning only valid integers by 'x|0' them.

Issue: http://code.google.com/p/google-web-toolkit/issues/detail?id=6130
Rietveld: http://gwt-code-reviews.appspot.com/1815803

+++ user/src/com/google/gwt/dom/client/DOMImplStandard.java
@@ -42,8 +42,8 @@ abstract class DOMImplStandard extends DOMImpl {
 
   @Override
   public native NativeEvent createMouseEvent(Document doc, String type,
-      boolean canBubble, boolean cancelable, int detail, int screenX,
-      int screenY, int clientX, int clientY, boolean ctrlKey, boolean altKey,
+      boolean canBubble, boolean cancelable, int detail, double screenX,
+      double screenY, double clientX, double clientY, boolean ctrlKey, boolean altKey,
       boolean shiftKey, boolean metaKey, int button, Element relatedTarget) /*-{
     // Because Event.getButton() returns bitfield values [1, 4, 2] for [left,
     // middle, right], we need to translate them to the standard [0, 1, 2]
